Organize serializers into a separate folder and start using controllers for gathering data
